# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "Auth_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Login successful — access tokens set in cookies",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request — validation failed or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — invalid email or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Authenticate user and obtain access tokens",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "Auth_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Token refreshed — new access token set in cookie",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — refresh token missing or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Refresh access token using refresh token cookie",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "Auth_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logout successful — cookies cleared",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — valid access token required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Log out and clear access/refresh token cookies",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/profile": {
      "get": {
        "operationId": "Auth_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Returns the authenticated user profile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — valid access token required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Get authenticated user profile",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Prescription Management API",
    "description": "API documentation for the MVP Prescription Management System.",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "doctor@clinic.com",
            "description": "User email"
          },
          "password": {
            "type": "string",
            "example": "***REDACTED-DEV-PASSWORD***",
            "description": "User password"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "Role": {
        "type": "string",
        "enum": [
          "ADMIN",
          "DOCTOR",
          "PATIENT"
        ],
        "description": "User role for RBAC"
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User unique identifier",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "doctor@clinic.com",
            "format": "email"
          },
          "role": {
            "description": "User role for RBAC",
            "example": "DOCTOR",
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ]
          }
        },
        "required": [
          "id",
          "email",
          "role"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Login successful",
            "description": "Confirmation message"
          },
          "user": {
            "description": "Authenticated user profile",
            "allOf": [
              {
                "$ref": "#/components/schemas/AuthUserDto"
              }
            ]
          }
        },
        "required": [
          "message",
          "user"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "HTTP status code of the error"
          },
          "message": {
            "type": "string",
            "example": "Invalid JSON payload",
            "description": "Error message describing what went wrong"
          },
          "path": {
            "type": "string",
            "example": "/auth/login",
            "description": "The request path that caused the error"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-05-13T10:30:00.000Z",
            "description": "ISO 8601 timestamp when the error occurred"
          }
        },
        "required": [
          "statusCode",
          "message",
          "path",
          "timestamp"
        ]
      },
      "RefreshResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Token refreshed",
            "description": "Confirmation message"
          }
        },
        "required": [
          "message"
        ]
      },
      "LogoutResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Logged out successfully",
            "description": "Confirmation message"
          }
        },
        "required": [
          "message"
        ]
      },
      "ThemePreference": {
        "type": "string",
        "enum": [
          "SYSTEM",
          "LIGHT",
          "DARK"
        ],
        "description": "UI theme preference"
      },
      "DoctorProfileSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "specialty": {
            "type": "string",
            "example": "Cardiology",
            "nullable": true
          },
          "medicalId": {
            "type": "string",
            "example": "MED-12345",
            "nullable": true
          },
          "signatureText": {
            "type": "string",
            "example": "Dr. Jane Doe",
            "nullable": true
          },
          "signatureImageUrl": {
            "type": "string",
            "example": "https://cdn.clinic.com/signatures/jane-doe.png",
            "format": "uri",
            "nullable": true
          }
        },
        "required": [
          "id"
        ]
      },
      "PatientProfileSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "birthDate": {
            "type": "string",
            "example": "1990-05-21T00:00:00.000Z",
            "format": "date-time",
            "nullable": true
          }
        },
        "required": [
          "id"
        ]
      },
      "UserProfileResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique user identifier (UUID v4)",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "doctor@clinic.com",
            "format": "email"
          },
          "role": {
            "description": "User role for RBAC",
            "example": "DOCTOR",
            "allOf": [
              {
                "$ref": "#/components/schemas/Role"
              }
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "ISO 8601 timestamp when the user was created",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "ISO 8601 timestamp when the user was last updated",
            "example": "2026-01-15T12:30:00.000Z"
          },
          "themePreference": {
            "description": "UI theme preference",
            "example": "SYSTEM",
            "allOf": [
              {
                "$ref": "#/components/schemas/ThemePreference"
              }
            ]
          },
          "doctor": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/DoctorProfileSummary"
              }
            ]
          },
          "patient": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/PatientProfileSummary"
              }
            ]
          }
        },
        "required": [
          "id",
          "email",
          "role",
          "createdAt",
          "updatedAt",
          "themePreference"
        ]
      },
...
    }
  }
}